digitap_backend/requirement.txt
arc4==0.0.2
asgiref==3.2.3
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
configobj==5.0.6
constantly==15.1.0
crypto==1.4.1
cryptography==2.8
cssselect==1.1.0
des==1.0.5
distro==1.4.0
Django==3.0.3
django-appconf==1.0.3
django-cors-headers==3.2.1
django-cryptography==1.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
Flask==1.1.1
gcloud==0.17.0
get==2019.4.13
googleapis-common-protos==1.51.0
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jws==0.1.3
lxml==4.4.2
M2CryptoWin32==0.21.1.post3
Markdown==3.2.1
MarkupSafe==1.1.1
mysql-client==0.0.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
Naked==0.1.31
numpy==1.18.1
oauth2client==3.0.0
pandas==1.0.1
parsel==1.5.2
Pillow==7.0.0
polling==0.3.1
post==2019.4.13
Protego==0.1.16
protobuf==3.6.1
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.4.3
pycryptodomex==3.4.3
PyDispatcher==2.0.5
PyHamcrest==1.10.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
Pyrebase==3.0.27
python-dateutil==2.8.1
python-dispatch==0.1.2
python-git==2018.2.1
python-jwt==2.0.1
python-scrapyd-api==2.1.2
pytz==2019.3
PyYAML==5.3
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
Scrapy==1.6.0
scrapyd==1.2.1
Send2Trash==1.5.0
service-identity==18.1.0
shellescape==3.8.1
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
tabula-py==2.0.4
Twisted==19.10.0
urllib3==1.25.8
vim-pck==0.1.0
w3lib==1.21.0
Werkzeug==1.0.0
zope.interface==4.7.1
